requirements.txt
# Core dependencies (always required)
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
asyncio>=3.4.3
aiohttp>=3.9.0
chromium-driver>=3.25.0
webdriver-manager>=4.0.0
psutil>=5.9.0
rich>=13.7.0
pydantic>=2.5.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
flask>=2.3.0
PyYAML>=6.0.0

# Stagehand AI browser automation (optional but recommended for AI-enhanced testing)
# Install separately with: npm install @browserbase/stagehand
# Python wrapper: pip install stagehand-python (if available)

# AI dependencies (optional - only needed when --ai flag is used)
# These can be commented out for no-AI deployments
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
groq>=0.4.0

# Testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
